[fix] 공유 썸네일 삭제 시 예외 추가 :  저장된 share thumbnail image가 있는 경우에만 s3에 삭제 요청

+++ src/main/java/jun/invitation/domain/shareThumbnail/service/ShareThumbnailService.java
@@ -48,7 +48,9 @@ public class ShareThumbnailService {
     public void deleteS3Image(ShareThumbnail shareThumbnail) {
         if (shareThumbnail != null) {
             String imageStoreFileName = shareThumbnail.getImageStoreFileName();
-            s3UploadService.delete(imageStoreFileName);
+            if (imageStoreFileName != null) {
+                s3UploadService.delete(imageStoreFileName);
+            }
         }
     }
